[ft2-user-sign-up-166204738]added controllers.
